+++ src/soorgeon/cli.py
@@ -12,19 +12,29 @@ def cli():
 @cli.command()
 @click.argument('path', type=click.Path(exists=True))
 @click.option('--log', '-l', default=None)
+@click.option(
+    '--df-format',
+    '-d',
+    default=None,
+    type=click.Choice(('parquet', 'csv')),
+    help='Format for variables with the df prefix. Otherwise uses pickle')
 @click.option('--product-prefix',
               '-p',
               default=None,
               help='Prefix for all products')
-def refactor(path, log, product_prefix):
+def refactor(path, log, product_prefix, df_format):
     """Refactor a monolithic notebook
     """
-    export.from_path(path, log, product_prefix=product_prefix)
+    export.from_path(path,
+                     log,
+                     product_prefix=product_prefix,
+                     df_format=df_format)
     click.secho(f'Finished refactoring {path!r}, use Ploomber to continue.',
                 fg='green')
+
     click.echo("""
-Install:
-    $ pip install ploomber
+Install dependencies (this will install ploomber):
+    $ pip install -r requirements.txt
 
 List tasks:
     $ ploomber status
@@ -32,7 +42,7 @@ List tasks:
 Execute pipeline:
     $ ploomber build
 
-Plot pipeline (note: this requires pygraphviz):
+Plot pipeline (this requires pygraphviz, which isn't installed by default):
     $ ploomber plot
 
 Documentation: https://docs.ploomber.io