+++ tests/test_proto.py
@@ -13,7 +13,8 @@ mixed_expected = '1 + 1 # Cell 1\n2 + 2 # Cell 3'
     [mixed, mixed_expected],
 ])
 def test_prototask_str(code, expected):
-    assert str(proto.ProtoTask('name', _read(code).cells)) == expected
+    assert str(proto.ProtoTask('name', _read(code).cells,
+                               df_format=None)) == expected
 
 
 @pytest.mark.parametrize('cells_idx, expected', [
@@ -22,8 +23,9 @@ def test_prototask_str(code, expected):
 def test_prototask_add_imports_cell(cells_idx, expected):
     cells = jupytext.reads(exploratory,
                            fmt='py:light').cells[cells_idx[0]:cells_idx[1]]
-    pt = proto.ProtoTask('task', cells)
+    pt = proto.ProtoTask('task', cells, df_format=None)
     cell = pt._add_imports_cell(exploratory,
                                 add_pathlib_and_pickle=False,
-                                definitions=None)
+                                definitions=None,
+                                df_format=None)
     assert cell['source'] == expected